Heuristic now returns a double

+++ src/Games/Kalaha/Players/Heuristic.java
@@ -7,5 +7,5 @@ import Games.Kalaha.Move;
  * Created by mrt on 19/04/16.
  */
 public interface Heuristic {
-     int compute(Board board);
+     double compute(Board board);
 }
